# pkg/apiserver/docs/swagger.yaml
definitions:
...
  AgentConditionStatus:
    enum:
    - "True"
    - "False"
    - Unknown
    type: string
    x-enum-varnames:
    - ConditionStatusTrue
    - ConditionStatusFalse
    - ConditionStatusUnknown
...
  Certificate:
    properties:
      apiVersion:
        description: APIVersion is the version of the API.
        type: string
      kind:
        description: Kind is the type of the resource.
        type: string
      metadata:
        allOf:
        - $ref: '#/definitions/CertificateMetadata'
        description: Metadata contains the metadata of the certificate.
      spec:
        allOf:
        - $ref: '#/definitions/CertificateSpec'
        description: Spec contains the specification of the certificate.
      status:
        allOf:
        - $ref: '#/definitions/CertificateStatus'
        description: Status contains the status of the certificate.
    type: object
  CertificateMetadata:
    properties:
      attributes:
        allOf:
        - $ref: '#/definitions/github_com_minuk-dev_opampcommander_api_v1.Attributes'
        description: Attributes are optional key-value pairs for the certificate.
      createdAt:
        description: CreatedAt is the timestamp when the certificate was created.
        type: string
      deletedAt:
        description: |-
          DeletedAt is the timestamp when the certificate was soft deleted.
          If nil, the certificate is not deleted.
        type: string
      name:
        description: Name is the unique name of the certificate.
        type: string
      namespace:
        description: Namespace is the namespace the certificate belongs to.
        type: string
    type: object
  CertificateSpec:
    properties:
      caCert:
        description: CaCert is the PEM-encoded CA certificate.
        type: string
      cert:
        description: Cert is the PEM-encoded certificate.
        type: string
      privateKey:
        description: PrivateKey is the PEM-encoded private key.
        type: string
    type: object
  CertificateStatus:
    properties:
      conditions:
        description: Conditions contains the conditions of the certificate.
        items:
          $ref: '#/definitions/Condition'
        type: array
    type: object
...
  Condition:
    properties:
      lastTransitionTime:
        type: string
      message:
        type: string
      reason:
        type: string
      status:
        $ref: '#/definitions/AgentConditionStatus'
      type:
        $ref: '#/definitions/ConditionType'
    type: object
  ConditionType:
    enum:
    - Created
    - Updated
    - Deleted
    - Connected
    - Healthy
    - Configured
    - Registered
    type: string
    x-enum-varnames:
    - ConditionTypeCreated
    - ConditionTypeUpdated
    - ConditionTypeDeleted
    - ConditionTypeConnected
    - ConditionTypeHealthy
    - ConditionTypeConfigured
    - ConditionTypeRegistered
...
  ListMeta:
    properties:
      continue:
        type: string
      remainingItemCount:
        type: integer
    type: object
...
  ListResponse-Certificate:
    properties:
      apiVersion:
        type: string
      items:
        items:
          $ref: '#/definitions/Certificate'
        type: array
      kind:
        type: string
      metadata:
        $ref: '#/definitions/ListMeta'
    type: object
...
  ListResponse-Namespace:
    properties:
      apiVersion:
        type: string
      items:
        items:
          $ref: '#/definitions/Namespace'
        type: array
      kind:
        type: string
      metadata:
        $ref: '#/definitions/ListMeta'
    type: object
...
  Namespace:
    properties:
      apiVersion:
        type: string
      kind:
        type: string
      metadata:
        $ref: '#/definitions/NamespaceMetadata'
      status:
        $ref: '#/definitions/NamespaceStatus'
    type: object
  NamespaceMetadata:
    properties:
      annotations:
        additionalProperties:
          type: string
        type: object
      createdAt:
        type: string
      deletedAt:
        type: string
      labels:
        additionalProperties:
          type: string
        type: object
      name:
        type: string
    type: object
  NamespaceStatus:
    properties:
      conditions:
        items:
          $ref: '#/definitions/Condition'
        type: array
    type: object
...
  github_com_minuk-dev_opampcommander_api_v1.Attributes:
    additionalProperties:
      type: string
    type: object
...
info:
  contact: {}
  description: This is the API server for OpAMP Commander, providing endpoints for
    managing OpAMP agents.
  termsOfService: http://swagger.io/terms/
  title: OpAMP Commander API Server
  version: "1.0"
paths:
...
  /api/v1/namespaces:
    get:
      description: Retrieve a list of namespaces.
      parameters:
      - description: Maximum number of namespaces to return
        in: query
        name: limit
        type: integer
      - description: Token to continue listing namespaces
        in: query
        name: continue
        type: string
      - description: Include soft-deleted namespaces
        in: query
        name: includeDeleted
        type: boolean
      - description: Label selector, e.g. env=prod,tier notin (canary,dev)
        in: query
        name: labelSelector
        type: string
      - description: Field selector (this resource exposes no selectable fields)
        in: query
        name: fieldSelector
        type: string
      - description: Case-sensitive name prefix filter
        in: query
        name: name
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ListResponse-Namespace'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List Namespaces
      tags:
      - namespace
...
  /api/v1/namespaces/{namespace}/certificates:
    get:
      description: Retrieve a list of certificates.
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Maximum number of certificates to return
        in: query
        name: limit
        type: integer
      - description: Token to continue listing certificates
        in: query
        name: continue
        type: string
      - description: Include soft-deleted certificates
        in: query
        name: includeDeleted
        type: boolean
      - description: Label selector, e.g. env=prod,tier notin (canary,dev)
        in: query
        name: labelSelector
        type: string
      - description: 'Field selector over the supported fields: metadata.namespace'
        in: query
        name: fieldSelector
        type: string
      - description: Case-sensitive name prefix filter
        in: query
        name: name
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/ListResponse-Certificate'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: List Certificates
      tags:
      - certificate
    post:
      consumes:
      - application/json
      description: Create a new certificate.
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Certificate to create
        in: body
        name: certificate
        required: true
        schema:
          $ref: '#/definitions/Certificate'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/Certificate'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create Certificate
      tags:
      - certificate
  /api/v1/namespaces/{namespace}/certificates/{name}:
    delete:
      description: Delete a certificate by its name.
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the certificate
        in: path
        name: name
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Delete Certificate
      tags:
      - certificate
    get:
      description: Retrieve a certificate by its name.
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the certificate
        in: path
        name: name
        required: true
        type: string
      - description: Include soft-deleted certificate
        in: query
        name: includeDeleted
        type: boolean
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Certificate'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Get Certificate
      tags:
      - certificate
    put:
      consumes:
      - application/json
      description: Update an existing certificate.
      parameters:
      - description: Namespace
        in: path
        name: namespace
        required: true
        type: string
      - description: Name of the certificate
        in: path
        name: name
        required: true
        type: string
      - description: Updated Certificate
        in: body
        name: certificate
        required: true
        schema:
          $ref: '#/definitions/Certificate'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/Certificate'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Update Certificate
      tags:
      - certificate
...
swagger: "2.0"